Stop unintentionally promoting a float to double

+++ code/qcommon/q_shared.h
@@ -537,7 +537,7 @@ static inline int VectorCompare(const vec3_t v1, const vec3_t v2)
 
 static inline vec_t VectorLength(const vec3_t v)
 {
-    return (vec_t)sqrt(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]);
+    return (vec_t)sqrtf(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]);
 }
 
 static inline vec_t VectorLengthSquared(const vec3_t v)
